fix: develop branch build오류 해결

See merge request s09-webmobile2-sub2/S09P12A409!22

+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/repository/BookmarkRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
     // insert는 기본 제공 save 사용
     
     // remove -> find로 찾아와서 isValid 0으로 변경
